+++ frontend/src/components/molecules/FeedbackButton.tsx
@@ -1,16 +1,9 @@
 import React, { useState } from 'react'
-import { usePostFeedback } from '../../services/api/feedback.hooks'
 import GTButton from '../atoms/buttons/GTButton'
-import GTModal from '../atoms/GTModal'
 import FeedbackView from '../views/FeedbackView'
 
 const FeedbackButton = () => {
     const [modalIsOpen, setModalIsOpen] = useState(false)
-    const [feedback, setFeedback] = useState('')
-    const { mutate: postFeedback } = usePostFeedback()
-    const submitFeedback = () => {
-        postFeedback({ feedback: feedback })
-    }
     return (
         <>
             <GTButton
@@ -19,17 +12,7 @@ const FeedbackButton = () => {
                 fitContent={false}
                 onClick={() => setModalIsOpen(true)}
             />
-            <GTModal
-                isOpen={modalIsOpen}
-                canClose={false}
-                onClose={() => setModalIsOpen(false)}
-                rightButtons={<GTButton onClick={submitFeedback} value="Send feedback" styleType="primary" />}
-                leftButtons={<GTButton onClick={() => setModalIsOpen(false)} value="Cancel" styleType="secondary" />}
-                title="Got Feedback?"
-                type="small"
-            >
-                <FeedbackView feedback={feedback} setFeedback={setFeedback} />
-            </GTModal>
+            <FeedbackView modalIsOpen={modalIsOpen} setModalIsOpen={setModalIsOpen} />
         </>
     )
 }
